fix #1595
git-svn-id: https://svn.modified-shop.org/svn/modified-shop/trunk@11680 79ac0487-9261-4f26-a34d-858dee0dcce0

+++ includes/classes/shipping.php
@@ -106,15 +106,15 @@
         $shipping_num_boxes = 1;
         $shipping_weight = $total_weight;
 
-        if (SHIPPING_BOX_WEIGHT >= $shipping_weight*SHIPPING_BOX_PADDING/100) {
-          $shipping_weight = $shipping_weight+SHIPPING_BOX_WEIGHT;
+        if ((double)SHIPPING_BOX_WEIGHT >= ($shipping_weight * (double)SHIPPING_BOX_PADDING / 100)) {
+          $shipping_weight = $shipping_weight + (double)SHIPPING_BOX_WEIGHT;
         } else {
-          $shipping_weight = $shipping_weight + ($shipping_weight*SHIPPING_BOX_PADDING/100);
+          $shipping_weight = $shipping_weight + ($shipping_weight * (double)SHIPPING_BOX_PADDING / 100);
         }
 
-        if (SHIPPING_MAX_WEIGHT != '' && $shipping_weight > SHIPPING_MAX_WEIGHT) { // Split into many boxes
-          $shipping_num_boxes = ceil($shipping_weight/SHIPPING_MAX_WEIGHT);
-          $shipping_weight = $shipping_weight/$shipping_num_boxes;
+        if ((double)SHIPPING_MAX_WEIGHT != '' && $shipping_weight > (double)SHIPPING_MAX_WEIGHT) { // Split into many boxes
+          $shipping_num_boxes = ceil($shipping_weight / (double)SHIPPING_MAX_WEIGHT);
+          $shipping_weight = $shipping_weight / $shipping_num_boxes;
         }
 
         $include_quotes = array();
